src/st_booking_api/api/v1/stadium/serializer.py
from rest_framework import serializers
from st_booking_api.stadium.models import Stadium, StadiumImage


class StadiumSerializer(serializers.ModelSerializer):
    class Meta:
        model = Stadium
        fields = '__all__'


class StadiumImageSerializer(serializers.ModelSerializer):
    class Meta:
        model = StadiumImage
        fields = '__all__'
