requirements.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
djoser==2.2.2
python-dotenv==1.0.1